// management/frontend/src/api/schemas/feedbackPublic.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Ego Collection Management
 * OpenAPI spec version: 0.1.0
 */

export interface FeedbackPublic {
  id?: string;
  created_at?: string | null;
  updated_at?: string | null;
  deleted_at?: string | null;
  creator_id?: string | null;
  updater_id?: string | null;
  deleter_id?: string | null;
  /**
     * @minLength 1
     * @maxLength 64
     */
  category: string;
  /**
     * @minLength 2
     * @maxLength 4000
     */
  content: string;
  /** @maxLength 128 */
  contact?: string;
  /** @maxLength 128 */
  submitter_username?: string;
  /** @maxLength 32 */
  status?: string;
  /** @maxLength 4000 */
  reply?: string;
  workspace_id: string;
  creator_name?: string | null;
  updater_name?: string | null;
}